Revert "Merge pull request #145 from PsychoLlama/feature/very-berry-yarn"
This reverts commit 0dd4101500484eaea03f14ea0461c7811254a281, reversing
changes made to 42befd2364be8189dec236e8e703173bef3e5b7d.

Reasoning: Yarn v2 was just an experiment to see what the experience was
like and how well it's integrating with the ecosystem. That's where it
failed for me. I use dependabot because spending the first hour of
a project updating everything is a huge pain. Apparently dependabot
doesn't actually use the `yarn` executable - it updates the lockfile
manually. That caused syntax errors because Yarn v2 changed their config
format to valid yaml.

In summary: I depend on dependabot, and Yarn v2 isn't compatible yet.

I'm reverting this until the landscape improves. Still, it was a fun
experiment, and definitely worth it for the sneak peak. The Yarn team is
doing some cool stuff.